Remove 'Sign that reads' from alt description of image
The page further down says
> Avoid using "Image of" or "Picture of"

The alt description of the example image read
'Sign that reads Warning do not read this sign'.
But that 'Sign that reads' is the same and as bad as 'Image of'.

If the image was a photo of a road with a sign on the side reading that,
then that would make more sense to say that it is a sign.
But in this example, it is irrelevant and not helpful
for screen reader users to know that it's supposed to be a sign.